Remove old cooking code (#1791)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Removes leftover code from the old cooking system, was making conflict
with the new cooking system.
This will also make it so the new sprites will be used.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Will fix #1477 as it was caused by the wrong cheese cake being called.

![image](https://github.com/user-attachments/assets/742c8f64-7e4d-4377-9185-37a80dbf6140)

Also cleans up a lot of commented out code.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->