Religious Clothing Adjustment (#1790)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Removes Eoran Acolyte "alternate" dress.

This is a server that aims for immersion, not promoting ERP to be
conducted in the church.

Acolytes of Eora are not prostitutes or bath wenches, they may have
"promiscious" mannerisms but they are not ladies of the night.

The code that allowed it may be preserved in this PR and could perhaps
be used on other clothing that is not religious in nature.


Additionally, fixes some minor issue mentioned regarding the Dendor
Amulet on Druids. For some reason the Dendor Amulet now uses the
psicross onmob. Until someone shows up with a proper Dendor onmob
sprite, it has been recolored to match the robes, so as not to look like
a "cum stain" on the druid robes.

![image](https://github.com/user-attachments/assets/95046f36-f694-4763-b2bc-3e65eb0954d7)