The return of alternative eoran robe
This PR is a compromise between the sides of those who want to remove the alternative robe and those who want to keep it. The person who removed the code but did not remove the alternative robes sprite felt it was not necessary to clean up the code. Thanks for that.

Now the eoran robe in the alternate does not show tits and crotch, which eliminates the possibility of ERP or whatever the creator of the PR hinted at about removing the alternative robe.

I hope we come to a consensus and don't remove everything we want just because we don't like it, though