EXPLOSIONS (#1112)
### W I P
## About The Pull Request

Reworks explosions as a prerequisite for reworked Fireballs

## Why It's Good For The Game

Will allow for Fun New Things, like explosion based fireballs that
aren't just Hotspot spawners

## Contents

**Explosion VisFX:** You can now add new visual effects to explosions
using the `visfx` flag in the `explosion()` proc. Add the new explosion
VFX to the list in explosion.dm
**Walls (can) Block Explosions:** Walls will take damage from explosions
without the blast dealing damage to the other side, unless it is a soft
material like a Tent.
**Better Projectile Explosions:** Projectiles like Fireball no longer
explode inside of walls (entirely), but rather backstep to the outside.
The wall still takes full damage by a mini-blast with the highest values
at 1 tile range.
**Damage based Turf destruction:** Explosions now use the same damage
system as weapons when it comes to damaging turfs. Bombs can blow up
walls without it being an RNG check for outright deleting them!
**Blast Mining:** Since rock is a turf, it also takes damage. Ores can
~~hopefully~~ be dropped by explosions, but provide ZERO Mining XP.

---------

Co-authored-by: dwasint <82520990+dwasint@users.noreply.github.com>