Architect Rebalance + Moved spawn position (#1771)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

- Architect SKILLS changed:
  - Mining 3 -> 1
  - Axes 0 -> 2
  - Carpentry 4 -> 3
  - Crafting 4 -> 3
  - Engineering 4 -> 2 (leave that stuff to the artificer)
  - Lumberjacking 1 -> 2
  - Blacksmithing 3 -> 2 (leave that stuff to the smith)
  - Reading 5 -> 4

- Architect STATS changed:
  - Strength +2 -> +1
  - Intelligence +4 -> +2
  - Endurance +2 -> +1

- Moves architect spawn to the now defunct mason's.

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Holy powercreep. Achitect is supposed to be the new and cooler mason,
not Leonardo Da Vinci.