Fixes keyrings not being able to use the keys. (#1812)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Makes the procs actually call the keys in the keyring for lockid.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Fixes an issue where interacting on a lockable thing with a keyring
wouldn't interact with it.


https://github.com/user-attachments/assets/57d68350-c2a4-4a2e-837a-d4d357e45d07

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->