Adds tunics to the crafting list (#1814)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Does as the title says, already tested. Never really knew why these were
exclusive to the merchant for some reason. Anyways more diversity on
clothing!
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

More clothes available for seamsters to make, and more choices for
people (When there's not a merchant around to buy it)
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->