Fixes faulds and battleskirts causing runtimes when put on belt (#1815)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

I am so sorry, i somehow tested it on pants and cloak but i forgot to
test it working on belts. If you put it on a belt it creates an
invisible fauld that you can't take off and it kind of... breaks things.
Again, sorry for missing this
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Bug fix! Again...
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->