Revert "Requested Sneak Changes & More" (#1806)
# about

Reverts Rotwood-Vale/Ratwood-Keep#1228

# why it's good

Apparently the PR, despite being intended as a sneak buff, has proven to
be a sneak nerf. From what others told me in the discord people can just
_see_ you if you're sneaking unless you're an expert sneaker... They
seemed to not like it. Suggesting a testmerge to see if it makes it
better.

We've still got the 'look around' button to detect sneakers.