Map Adjustments (#1826)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR makes a few map changes.
First of all the Lavamines is now actually woth mining at:

![image](https://github.com/user-attachments/assets/c4fb8e6a-63ed-4b1d-9e9c-e9d1a4c83d3c)

![image](https://github.com/user-attachments/assets/7863380b-3884-4d09-8aec-01424409ea6b)
Prior to this change, it was entirely randomized, and not worth mining
here, compared to other areas.

Secondly: There is no longer a mineable gap on the second level, on
either side of the vanguard gate. This was an oversight, according to
mappers, and has been corrected.
Thirdly: Some new minor buildings and things have been added. To quickly
list them and provide pictures:
A shack has been added within the vanguard gates, to replace the one
that has been missing since the major map change. (We all know what that
old shack was used for, this one simply re-adds it for the same purpose)

![image](https://github.com/user-attachments/assets/20457e41-e140-4eaf-919c-98d4b7c03c9f)
The chapel in the bog now has an astrata statue, and features a under
section dedicated to noc:

![image](https://github.com/user-attachments/assets/0ed3326a-21d6-45fa-a922-69aa82a4a552)

![image](https://github.com/user-attachments/assets/23562074-dfd4-48ff-baa4-7707ba84b57e)

And finally, a little cliffside shack outside the vanguard gate,
infested by deadites.

![image](https://github.com/user-attachments/assets/4505193a-f15e-4e54-ad3a-0f11a764601b)



<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
More reasons to go to the bog, Minor map additions, 
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->