Gender Equality or Something (#1840)
## About The Pull Request
Allows Confessors to be women, same as the Inquisitor. This seemed like
oversight on my part, from when I re-added it. Additionally, it permits
women to be the Sheriff.

## Why It's Good For The Game
something about rights i guess or whatever i 'unno
Sheriff was approved by Useroth. Confessor is more-or-less just
oversight, as noted. This is a correction to that oversight.