Bugfix: Restores #include for floorinteraction.dm (#1856)
## About The Pull Request

Slurping cum off the floor was broken by accident while porting the
cleanup from the not-smut server. It was overlooked that they purged
smut-code from their files. (In this one:
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/1660/files#diff-d998047e4b22ab7bcf2fbb8ac64f58dc648455a7399efcb28d8c25acf2ff771a
)

## Why It's Good For The Game

Vital for advanced maid gameplay. (For those curious, you need to lie
down, set bite intent and click the floor)


![image](https://github.com/user-attachments/assets/0986b14b-148f-42cb-93eb-1f1ec508ed3c)