Vamp undead faction and better telepathy message (#1863)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Gives vamp undead faction (because they are undead) so zombies and
skeletons wont attack them, which is annoying

![image](https://github.com/user-attachments/assets/5c9fcb68-5a05-4958-898e-7ccf9ea25f3f)
Also makes telepathy message a bit better

![image](https://github.com/user-attachments/assets/1a1cafec-3eec-4932-908e-c94791b541e4)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
it is not
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->