Makes soilsons able to craft peasant weapons (#1860)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Lowers craft diff for peasant weapons to 2, which is what the soilsons
get
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
"The soilers don't have enough craft to even make the improvised
weapons, which are weapons they're supposed to use. With Goblins in the
game now constantly raiding the Soilers 5 seconds into the round, having
a weapon to defend themselves would be good. The improvised weapons are
supposed to be used for Peasant Rebellions, they require way more
materials to make than regular weapons. In order to make an improvised
billhook, you need to smith a Sickle, get rope and a log and then make
the improvised billhook, when you could probably smith a regular
billhook instead. They're weapons that are designed in the game to be
made out of tools the Soilers already have, none of them can be stowed
on the Soiler and HAVE to be held onto. The Peasant Flail requires so
much strength most soilers can't use it either. They're not good, it
wouldn't be OP to make them actually able to be used by the people who
should use them"
Requested on discord.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->