[Port] Some QoL from Azure Peak again (Armour coverage examine, better colours for Rockhill (#1861)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

1. Added coverage to armor inspect text.
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/1008


![image](https://github.com/user-attachments/assets/e0eee76a-3fd7-452e-8481-1f07fb9b97f7)


2. "Adjusts the hexcodes of the heraldry colors, as well as those in the
colors.dm of roguetown, to fit better visually. No more piss-yellow
carpets in the Keep. Plus they're better on the eyes." -
[MortoSasye](https://github.com/MortoSasye)
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/995


![image](https://github.com/user-attachments/assets/8c42da0d-591a-46e3-b226-f6f872418613)


3. Speeds up bait deploying time to 2 seconds.
"So, apparently using baits probably takes as long as it does because in
OG roguecode there was a planned hunting skill that never saw the light
of day that presumably would influence the speed of deploying/retrieving
bait. This skill was never really implemented, so all we have are very
inconveniently lengthy bait usage times. This long usage time doesn't
really add anything to the gameplay experience at all and really
shouldn't be as long as it is anyway, so let's just speed it up." -
[Tenryumi](https://github.com/Tenryumi)
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/968



**_Also, someone deleted/broke an ability to examine armor durability
coifs. I'll try fix it soon_**