Adds A Per-Slot Virginity Toggle (#1878)
## About The Pull Request

I have seen multiple players ask for a way to toggle being a virgin.
This enables such. The toggle is per-character and is overridden by the
normal roles that override virginity. Namely duke/duchess _must_ be
non-virgins as they popped out the heirs, archivist/steward are default
virgins according to the code, and a good few church/clergy roles are
default virgin.

## Why It's Good For The Game

More customization good. Also multiple players want it.

## Proof Of Testing


![Screenshot 2024-12-11
034707](https://github.com/user-attachments/assets/6d0eabb9-edc6-4065-8a26-5241eff1caca)
![Screenshot 2024-12-11
035145](https://github.com/user-attachments/assets/b3387e39-8811-4220-b6ae-ce8f03aa521f)

---------

Co-authored-by: Heheniy <turodore@gmail.com>