Adds clock to the bell (#1889)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Well many people have asked for one and nobody seems to have delivered,
so i will. Adds a clock to the bell! Now you can go ahead and hit that
thing right on the dot.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Probably gonna make the bell be used more, and it's nice to have
something such as someone ringing the bell at 12:00 just to inform the
people about the time. Or for other uses, of course.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->