Hoardmaster items (#1901)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
- Horned cap + Gorget added to Mage and Sawbone
- Price adjusted for the studded leather armor, so it cost the same as
in the sawbone tab
- Reinforced hood was moved so people could RP while being armored
- Most of all the raincoats were added
- Fur coat
- Tabard
- And a half-mask
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
People like to meta the raincoats, now it's a bit harder
A basic helmet was added to the Mage and Sawbone so they did not have to
beg or rummage through other people's tab

A lot of clothes options, which are always nice.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->