Harbour District + Brothel & Nightmaster + Sailors + Naval Shipments mechanic (#1852)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## Points of interest
A brand new city district north of the church and west of the keep, the
harbor.
The district includes:
- Reworked mercenary guild.
- Brand-new brothel, run by the newly re-enabled Nightmaster and filled
with prostitutes.
- Eastern docks, with latejoin arrival spawns and the
longshoreman/harbormaster ship (defacto headquarters).
- Underground storage, accessible only by the sailors.
- Northern docks, where the shipments come and go.
- Cute lighthouse north-east.

## The Brothel
The brothel, also known as the Whitevein lounge, is a long decayed bath
house turned into a den of low lives. It is managed by the
nightmaster/nightmatron, with several prostitutes under his employment.

## The sailors
The harbor is managed by a group of longshoremen, headed by a
harbormaster, a retired captain. They handle the flow of goods and are
often in direct contact with the merchant, who signs shipment orders.

## Shipments
The boat can be called by ringing the bell at the northern docks. Anyone
can then leave goods on it, send it back and it will return with the
money. Leaving a shipment order with the needed money will result in the
order being delivered on the next pass.

**Shipment orders can only be made by the merchant, by using an empty
papyrus scroll on the Ledger (it spawns on the table next to his bed)**

(I will make steward imports also come through the ship eventually)


## Credits
- (working shipments)
https://github.com/Vanderlin-Tales-Of-Wine/Vanderlin/pull/253
- (working shipments)
https://github.com/Vanderlin-Tales-Of-Wine/Vanderlin/pull/46.
- (ship design) https://github.com/Rotwood-Vale/Ratwood-Keep/pull/114.
- (harbormaster / longshoreman roles / sand sprite)
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/205.
- Everything else - Gyran, Nikolay, Izzy.

---------

Co-authored-by: dwasint <82520990+dwasint@users.noreply.github.com>
Co-authored-by: ni.ushakov <strider48954@gmail.com>
Co-authored-by: Heheniy <turodore@gmail.com>