SCOM Calls (#1904)
Adds SCOM Jabberlines! Examine a SCOM to see its number designation. MMB
the SCOM to put in the number of the SCOM you want to call. LMB to
accept, RMB to reject. When in a call, the speech pickup radius is no
longer limited to right in front of the SCOM. LMB to mute the
microphone, RMB to mute the speaker. MMB to cut the jabberline.


![image](https://github.com/user-attachments/assets/8caec449-cf9e-4505-9302-6a111e76370b)

You will need to stand in front of a SCOM to actually read the laws
list.

Also makes it so that Shylocks give you round numbers for your account
balance.