Sea Jackets for Longshoremen (#1909)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Replaces the leather armor on longshoremen with sea jackets, adds a
bottle of wine like how the gambler job gets to their satchel.

## Why It's Good For The Game

Fits better thematically, looks so much better aesthetically, and
functions nearly the same. If longshoremen want to go battle, they're
going to go ask the seamstress for better reinforced leather or go to
the smithy regardless. Also the sea jacket has a pouch to put the bottle
of wine you start with your satchel into. Because who doesn't love a
drunken sailor early in the morning?