Buffs Harp Song Volume, Adds 4 New Songs (#1911)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This PR increases the volume of the current harp songs as much as I
could through audacity. They are slightly more noticeable now. I also
added 4 new songs which are royalty free from soundcloud, each
thematically named. These songs are good quality too and are pure harp,
much more noticeable than the old songs. The harp just might be a new
favorite instrument.

Also I edited the guitar and harp description a bit because the guitar
description was dumb and the harp was too simple.

## Why It's Good For The Game

Currently, the harp is very underused. Nightmaidens will just throw the
harp out, and I do too because the music sucks and you can't hear shit
without pumping your volume up, but zaping your ears from other sounds
when doing so. Now I'll actually want to keep it on me instead of
tossing it on the bathouse table and bumrushing the inn for the lute.
The new songs also add some atmosphere, moreso than the old ones.

Here are the songs I added:
https://soundcloud.com/sei_peridot/wish3_harponly
https://soundcloud.com/sei_peridot/wish4-harponly
https://soundcloud.com/sei_peridot/wish5_harponly
https://soundcloud.com/sei_peridot/minstrel2_harp