Archivist (Scholar) Overhaul (#1905)
This update introduces a rework of the archivist, making it the scholar
and adding a brand new feature to the game: teaching.


![image](https://github.com/user-attachments/assets/8ac870db-c627-439f-8bb9-96c091999aa8)

Archivist has always been a neglected role, so I believed it must've
been made actually useful. The scholar has better skills and more
abilities, including multiple spells such as prestidigitation, fetch and
message. They are also able to learn new spells, though for the sake of
balance they start with 0 spell points.


![image](https://github.com/user-attachments/assets/275f1a31-8f14-49f3-b397-7b36f6f58279)

And now the best part... Teaching! The scholar is able to help other
player improve their picked skill once, granting this role a unique
ability.


![image](https://github.com/user-attachments/assets/06564c22-d0a8-4b50-8232-5b4dc381d4f2)

![image](https://github.com/user-attachments/assets/571f904b-b22d-4871-a291-369209ee208e)