Silent casting fix
I have no idea how tracking mouse clicks works here or how the spell system works.

The original issue is that you can, via repeated clicks, bypass into:
/obj/effect/proc_holder/spell/invoked/InterceptClickOn
by attempting to cast repeatedly. Once it's there, neither of the conditional checking functions in:
 can_cast
 cast_check

check for how long the spell is casting nor whether the spell is set to not be released early so the spell simply insta casts, This fixes it with some caveats.
 Firstly, the spell icons have a chance to be set to be on cooldown if you cancel your cast after starting it. I suspect there's a flag in a function I'm missing

Somehow the 'doneset' variable accurately tracks the min-cast time of the spell that's being cast and when it's up, the spells is fully charged and can be safely cast. However, the entire function is called by various hud elements

'curplaying' also seems to be an important variable.

There is a more elegant solution but I can't think of it right now.