Adds rings to treasury bounty, increases default bounty payout, and adds satchel to steward (#1918)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Title

## Why It's Good For The Game

The treasury bounty is an underused mechanic. What it does is pay people
in their accounts a % of the value of certain items, with this PR the
list of items to pay becoming statues, rings, gems, and cups. Also, the
PR adds a satchel to the steward's default gear because, for some
reason, the clerk started with it and they didn't.