Herald Invulnerability Removed
Because of a newer exploit, and how easy it is to ascend the Cult otherwise, I feel as if this isn't as important. Others might disagree, but it's probably wise to go back to how it used to function.

That is, removing its literal godmode.
This also provides it with dodge expert, to alleviate some of the issue of it being grabbed and held in place to prevent actually fighting it. To some degree, at least. Hopefully. Maybe.