update_filter runtime fix
When you have two or more filters applied to a character, it would case a runtime error because the function in atoms wanted it to be sorted. The comparator used was default. However, in the case of filters, it was passed two lists.

Added a comparator to check the 'priority' key of those lists.