Herald Invulnerability Removed (#1925)
## About The Pull Request
Because of a newer exploit, and how easy it is to ascend the Cult
otherwise, I feel as if it isn't as important to have the herald
invincible. Players can once again form groups of like twelve to take it
down, as they did prior. Others might disagree, but it's probably wise
to go back to how it used to function.

That is, removing its literal godmode.

## Why It's Good For The Game
I 'unno.