Town square update and scholar expansion (#1929)
Changes the town square to create a better atmosphere in town, also
fixes a bug making the scholar spawn in the keep's archives instead of
the academy.


![StrongDMM_qSdvauBqGb](https://github.com/user-attachments/assets/9dfb6867-61de-4c88-9dcb-82b579429ef0)

![StrongDMM_tY7To5xZUX](https://github.com/user-attachments/assets/ad03b19b-140d-4d19-b43d-58a4db24e17d)

![StrongDMM_Hub9fwZsti](https://github.com/user-attachments/assets/1b9cd587-b7a1-4433-8492-bbe6e1901eb6)

The scholar also got an update and now has a laboratory to brew potions
to sell them. Since the alchemist is disabled, it'd be great to have a
potion merchant, it also adds more content for the scholar than just
teaching.


![image](https://github.com/user-attachments/assets/c6f74289-742d-4e72-8451-41087bf6f0a6)