Tin Expansion Part 1: Cutlery (#1932)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Tin is a metal currently second to none in being useless, and this PR
aims to change that by introducing tin-based cutlery in the form of
Pewterware, an alloy made mostly of tin.

Smiths can now make mildly valuable Pewter cutlery from tin ingots --
Which, while not very expensive by any means, allows idle smiths to
produce something to sell to the merchant/steward for cash in their
spare time.

This PR also adds silverware cutlery. I will map it to the keep in a
later PR. The recipes are also added. Silverware is a much more
expensive variant of Pewter.

Figure 1: Pewterware and Silverware added by this PR:


![Capture2](https://github.com/user-attachments/assets/551c42ac-325c-4310-9349-11baf1ad1a3f)

Figure 2: Pewterware and Silverware sprites:


![Capture](https://github.com/user-attachments/assets/2778a3fd-9731-45c2-98ea-49da55e29a5c)

Sprites are currently hue-shifted placeholders, except for platters,
which are OC but still placeholder sprites.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

This allows smiths to work towards items that can be reliably sold and
exported to a merchant.

Later on, this will also allow for valuable silverware to be added to
the keep -- And less valuable Pewterware to be added to the town's
houses, both making the game more immersive and giving beggars/thief
players more incentive to try and steal things -- While also not being
so expensive that they are a meta-rush.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->